+++ fn/core.go
@@ -1,6 +1,7 @@
 package fn
 
 import (
+	"cmp"
 	"slices"
 	"strings"
 
@@ -16,5 +17,29 @@ func CoreSortedPartition[T any](values []discrete.Value, items []T) discrete.Sol
 		partition := sortedPartitionGroups(groups)
 		slices.Sort(partition)
 		return strings.Join(partition, "/")
+	}
+}
+
+// CoreMirroredSequence groups the normal and the reversed sequence as one
+func CoreMirroredSequence[T any](items []T) discrete.SolutionCoreFn {
+	return func(solution *discrete.Solution) string {
+		sequence := sequenceStrings(solution, items)
+		first, last := sequence[0], list.Last(sequence, 1)
+		if cmp.Compare(first, last) == 1 {
+			slices.Reverse(sequence)
+		}
+		return strings.Join(sequence, " ")
+	}
+}
+
+// CoreMirroredValues groups the normal and reversed list of values as one
+func CoreMirroredValues[T any](p *discrete.Problem, items []T) discrete.SolutionCoreFn {
+	return func(solution *discrete.Solution) string {
+		output := valueStrings(p, solution, items)
+		first, last := output[0], list.Last(output, 1)
+		if cmp.Compare(first, last) == 1 {
+			slices.Reverse(output)
+		}
+		return strings.Join(output, " ")
 	}
 }